pyproject.toml
[project]
name = "grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.68.1",
    "grpcio-tools>=1.68.1",
]
